+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## [Unreleased]
 
+- Add `Inode::dtime_val` and `Inode::set_dtime_val` for handling of deleted inodes.
 - Dependency bump
 - More fsck fixes
     - Include fix for bg group bitmap checksum calculation
     - Also fixes superblock not being written back properly
-    - 
 
 ## 0.1.0-alpha.7